Detect whether firmware has already been uploaded this boot or not.
When the NUI Audio device is running the bootloader, it has one USB interface.
When the NUI Audio device is running the uploaded firmware, it has a second one
(which does not appear to be used).

Thanks to Sebastian Ortiz for noticing the difference in the configuration
descriptor that makes this detection possible.

Signed-off-by: Drew Fisher <drew.m.fisher@gmail.com>